requirements.txt
openai
qdrant-client
sentence-transformers
python-dotenv
